# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dental Case Management System API",
    "description": "نظام إدارة حالات الأسنان - واجهة البرمجة",
    "version": "1.0.0",
    "contact": {
      "name": "Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "secretary", "designer", "finisher"]
          },
          "status": {
            "type": "string",
            "enum": ["online", "offline", "idle"]
          },
          "department": {
            "type": "string"
          }
        }
      },
      "DentalCase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "caseNumber": {
            "type": "string"
          },
          "patientName": {
            "type": "string"
          },
          "patientEmail": {
            "type": "string"
          },
          "patientPhone": {
            "type": "string"
          },
          "currentStage": {
            "type": "string",
            "enum": ["waiting", "secretary", "design", "khart", "finishing", "completed"]
          },
          "status": {
            "type": "string",
            "enum": ["waiting", "in_progress", "completed"]
          },
          "priority": {
            "type": "string",
            "enum": ["low", "normal", "high", "urgent"]
          },
          "assignedTo": {
            "$ref": "#/components/schemas/User"
          },
          "createdBy": {
            "$ref": "#/components/schemas/User"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "caseId": {
            "type": "string"
          },
          "caseNumber": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "enum": ["created", "assigned", "reassigned", "moved_stage", "completed", "reopened", "released"]
          },
          "performedBy": {
            "$ref": "#/components/schemas/User"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "details": {
            "type": "object"
          }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Server is running"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "تسجيل الدخول",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cases": {
      "get": {
        "summary": "الحصول على جميع الحالات",
        "tags": ["Cases"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cases"
          }
        }
      },
      "post": {
        "summary": "إنشاء حالة جديدة",
        "tags": ["Cases"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "patientName": {
                    "type": "string"
                  },
                  "patientEmail": {
                    "type": "string"
                  },
                  "patientPhone": {
                    "type": "string"
                  },
                  "caseType": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string"
                  },
                  "dueDate": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Case created successfully"
          }
        }
      }
    },
    "/api/cases/{id}/claim": {
      "put": {
        "summary": "استدعاء حالة",
        "tags": ["Cases"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case claimed successfully"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "الحصول على جميع المستخدمين",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        }
      }
    },
    "/api/audit-logs": {
      "get": {
        "summary": "الحصول على سجلات العمليات",
        "tags": ["Audit"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of audit logs"
          }
        }
      }
    },
    "/api/notifications": {
      "get": {
        "summary": "الحصول على التنبيهات",
        "tags": ["Notifications"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications"
          }
        }
      }
    }
  }
}